+++ packages/webapp-libs/webapp-core/src/components/forms/form/formControl/formControl.component.tsx
@@ -1,7 +1,7 @@
 import { Slot } from '@radix-ui/react-slot';
 import * as React from 'react';
 
-import { useFormField } from '../formField';
+import { useFormField } from '../';
 
 const FormControl = React.forwardRef<React.ElementRef<typeof Slot>, React.ComponentPropsWithoutRef<typeof Slot>>(
   ({ ...props }, ref) => {
