+++ packages/webapp-libs/webapp-core/src/components/forms/form/formField/formField.component.tsx
@@ -1,8 +1,6 @@
 import * as React from 'react';
 import { Controller, ControllerProps, FieldPath, FieldValues, useFormContext } from 'react-hook-form';
 
-import { FormItemContext } from '../formItem';
-
 type FormFieldContextValue<
   TFieldValues extends FieldValues = FieldValues,
   TName extends FieldPath<TFieldValues> = FieldPath<TFieldValues>
@@ -25,27 +23,4 @@ const FormField = <
   );
 };
 
-const useFormField = () => {
-  const fieldContext = React.useContext(FormFieldContext);
-  const itemContext = React.useContext(FormItemContext);
-  const { getFieldState, formState } = useFormContext();
-
-  const fieldState = getFieldState(fieldContext.name, formState);
-
-  if (!fieldContext) {
-    throw new Error('useFormField should be used within <FormField>');
-  }
-
-  const { id } = itemContext;
-
-  return {
-    id,
-    name: fieldContext.name,
-    formItemId: `${id}-form-item`,
-    formDescriptionId: `${id}-form-item-description`,
-    formMessageId: `${id}-form-item-message`,
-    ...fieldState,
-  };
-};
-
-export { useFormField, FormField };
+export { FormField, FormFieldContext };
