+++ packages/webapp-libs/webapp-core/src/components/forms/form/index.ts
@@ -5,3 +5,4 @@ export * from './formField';
 export * from './formItem';
 export * from './formLabel';
 export * from './formMessage';
+export * from './useFormField';
