+++ packages/webapp-libs/webapp-core/src/components/forms/form/useFormField/useFormField.hook.ts
@@ -1,0 +1,29 @@
+import React from 'react';
+import { useFormContext } from 'react-hook-form';
+
+import { FormFieldContext, FormItemContext } from '../';
+
+const useFormField = () => {
+  const fieldContext = React.useContext(FormFieldContext);
+  const itemContext = React.useContext(FormItemContext);
+  const { getFieldState, formState } = useFormContext();
+
+  const fieldState = getFieldState(fieldContext.name, formState);
+
+  if (!fieldContext) {
+    throw new Error('useFormField should be used within <FormField>');
+  }
+
+  const { id } = itemContext;
+
+  return {
+    id,
+    name: fieldContext.name,
+    formItemId: `${id}-form-item`,
+    formDescriptionId: `${id}-form-item-description`,
+    formMessageId: `${id}-form-item-message`,
+    ...fieldState,
+  };
+};
+
+export { useFormField };
